feat(authorize admin): controle de Update e Delete pela Role Admin
Os personagens e Filmes originais da franquia só podem ser alterados os excluídos por usuários com a
role admin

+++ FandomStarWars.API/Controllers/MovieController.cs
@@ -3,7 +3,7 @@ using FandomStarWars.Application.DTO_s;
 using FandomStarWars.Application.Interfaces;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-//using Refit;
+using System.Security.Claims;
 
 namespace FandomStarWars.API.Controllers
 {
@@ -12,7 +12,9 @@ namespace FandomStarWars.API.Controllers
     public class MovieController : ControllerBase
     {
         private readonly IMovieService _movieService;
-        
+        private const int IdLastMovieOrigin = 6;
+
+
         public MovieController(IMovieService movieService)
         {
             _movieService = movieService;
@@ -56,6 +58,17 @@ namespace FandomStarWars.API.Controllers
         [Authorize]
         public async Task<ActionResult<MovieDTO>> Update(MovieDTO movie)
         {
+            var role = HttpContext.User.FindFirstValue(ClaimTypes.Role);
+
+            if (movie.Id <= IdLastMovieOrigin && role is not "Admin")
+            {
+                return StatusCode(403, new GenericResponse
+                {
+                    IsSuccessful = false,
+                    Message = "Ops, parece que você não tem permissão de alterar este filme, confira se possua uma conta Admin"
+                });
+            }
+
             var response = _movieService.UpdateAsync(movie).Result;
             return response.IsSuccessful ? Ok(response) : BadRequest(response);
         }
@@ -65,6 +78,17 @@ namespace FandomStarWars.API.Controllers
         [Route("{id}")]
         public async Task<ActionResult<MovieDTO>> Delete(int id)
         {
+            var role = HttpContext.User.FindFirstValue(ClaimTypes.Role);
+
+            if (id <= IdLastMovieOrigin && role is not "Admin")
+            {
+                return StatusCode(403, new GenericResponse
+                {
+                    IsSuccessful = false,
+                    Message = "Ops, parece que você não tem permissão de deletar este filme, confira se possua uma conta Admin"
+                });
+            }
+
             var response = _movieService.DeleteAsync(id).Result;
             return response.IsSuccessful ? Ok(response) : BadRequest(response); 
         }
